+++ src/components/ids-data-grid/ids-data-grid-header.ts
@@ -94,6 +94,7 @@ export default class IdsDataGridHeader extends IdsEventsMixin(IdsElement) {
     let x = 0;
     let w = 0;
     let columnId = '';
+    let columnElem: HTMLElement;
 
     const mouseMoveHandler = (e: MouseEvent) => {
       // Determine how far the mouse has been moved
@@ -109,7 +110,8 @@ export default class IdsDataGridHeader extends IdsEventsMixin(IdsElement) {
 
       this?.style.setProperty('cursor', '');
       requestAnimationFrame(() => {
-        this.dataGrid!.isResizing = false;
+        this.dataGrid.isResizing = false;
+        columnElem?.parentElement?.querySelectorAll('[draggable]').forEach((el: Element) => el.setAttribute('draggable', 'true'));
       });
     };
 
@@ -125,9 +127,11 @@ export default class IdsDataGridHeader extends IdsEventsMixin(IdsElement) {
       x = e.clientX;
 
       // Calculate the current width of column
-      const col = target.closest('.ids-data-grid-header-cell');
-      const colStyles = window.getComputedStyle(col);
-      columnId = col.getAttribute('column-id');
+      columnElem = target.closest('.ids-data-grid-header-cell') as HTMLElement;
+      columnElem.parentElement?.querySelectorAll('[draggable]').forEach((el) => el.setAttribute('draggable', 'false'));
+
+      const colStyles = window.getComputedStyle(columnElem);
+      columnId = columnElem.getAttribute('column-id') || '';
       w = parseInt(colStyles.width, 10);
 
       // Attach listeners for document's events
@@ -138,7 +142,7 @@ export default class IdsDataGridHeader extends IdsEventsMixin(IdsElement) {
       this?.style.setProperty('cursor', 'col-resize');
 
       // Prevent a click causing a sort
-      this.dataGrid!.isResizing = true;
+      this.dataGrid.isResizing = true;
     });
   }
 
@@ -153,32 +157,24 @@ export default class IdsDataGridHeader extends IdsEventsMixin(IdsElement) {
     let dragInitiated = false;
 
     // Style the Dragger
-    this.offEvent('dragstart.resize', this);
-    this.onEvent('dragstart.resize', this, (e: DragEvent) => {
+    this.offEvent('dragstart.reorder', this);
+    this.onEvent('dragstart.reorder', this, (e: DragEvent) => {
+      if (this.dataGrid.isResizing) return;
       const target = (e.target as any);
-      if (!target.classList.contains('reorderer')) {
-        return;
-      }
-
       dragInitiated = true;
-      target.parentNode.classList.add('active-drag-column');
-      dragger = target.parentNode.cloneNode(true);
+      target.classList.add('active-drag-column');
+      dragger = target.cloneNode(true);
       dragger.classList.add('dragging');
-      dragger.style.position = 'absolute';
-      dragger.style.top = '0';
-      dragger.style.left = '-1000px';
 
       this?.appendChild(dragger);
       // Based on width of 110
       e?.dataTransfer?.setDragImage(dragger, this.dataGrid?.localeAPI.isRTL() ? 100 : 10, 18);
-
-      target.style.position = 'absolute';
-      startIndex = target.parentNode.getAttribute('aria-colindex');
+      startIndex = target.getAttribute('aria-colindex');
     });
 
     // Show the arrows
-    this.offEvent('dragenter.resize', this);
-    this.onEvent('dragenter.resize', this, (e: DragEvent) => {
+    this.offEvent('dragenter.reorder', this);
+    this.onEvent('dragenter.reorder', this, (e: DragEvent) => {
       if (!dragInitiated) {
         // Accept Dropped Text
         if ((e.target as HTMLElement)?.getAttribute('color-variant') === 'alternate-formatter') {
@@ -199,22 +195,22 @@ export default class IdsDataGridHeader extends IdsEventsMixin(IdsElement) {
           - (this.offsetParent as HTMLElement).offsetLeft;
       }
       dragArrows?.style.setProperty('left', `${this.dataGrid?.localeAPI.isRTL() ? cellRight - offsetLeft : cellLeft - offsetLeft}px`);
-      dragArrows?.style.setProperty('height', `${rect.height}px`);
+      dragArrows?.style.setProperty('height', `${rect.height - 1}px`);
       dragArrows?.style.setProperty('display', 'block');
 
       e.preventDefault();
     });
 
     // Use a normal cursor (not drag and drop)
-    this.offEvent('dragover.resize', this);
-    this.onEvent('dragover.resize', this, (e: DragEvent) => {
+    this.offEvent('dragover.reorder', this);
+    this.onEvent('dragover.reorder', this, (e: DragEvent) => {
       if (!dragInitiated) return;
       e.dataTransfer!.dropEffect = 'move';
       e.preventDefault();
     });
 
-    this.offEvent('dragleave.resize', this);
-    this.onEvent('dragleave.resize', this, (e: DragEvent) => {
+    this.offEvent('dragleave.reorder', this);
+    this.onEvent('dragleave.reorder', this, (e: DragEvent) => {
       if (!dragInitiated && (e.target as HTMLElement)?.getAttribute('color-variant') === 'alternate-formatter') {
         (e.target as any).style.backgroundColor = '';
       }
@@ -228,8 +224,8 @@ export default class IdsDataGridHeader extends IdsEventsMixin(IdsElement) {
     };
 
     // Set everything temp element back to normal
-    this.offEvent('dragend.resize', this);
-    this.onEvent('dragend.resize', this, (e: DragEvent) => {
+    this.offEvent('dragend.reorder', this);
+    this.onEvent('dragend.reorder', this, (e: DragEvent) => {
       if (!dragInitiated) {
         if ((e.target as HTMLElement)?.getAttribute('color-variant') === 'alternate-formatter') {
           (e.target as any).style.backgroundColor = '';
@@ -240,8 +236,8 @@ export default class IdsDataGridHeader extends IdsEventsMixin(IdsElement) {
       removeDragger(e);
     });
 
-    this.offEvent('drop.resize', this);
-    this.onEvent('drop.resize', this, (e: DragEvent) => {
+    this.offEvent('drop.reorder', this);
+    this.onEvent('drop.reorder', this, (e: DragEvent) => {
       if (!dragInitiated) {
         if ((e.target as HTMLElement)?.getAttribute('color-variant') === 'alternate-formatter') {
           (e.target as any).style.backgroundColor = '';
@@ -359,7 +355,7 @@ export default class IdsDataGridHeader extends IdsEventsMixin(IdsElement) {
     `;
 
     const resizerTemplate = `<span class="resizer"></span>`;
-    const reorderTemplate = `<div class="reorderer" draggable="true"><ids-icon icon="drag" size="medium"></ids-icon></div>`;
+    const reorderTemplate = `<div class="reorderer"><ids-icon icon="drag" size="medium"></ids-icon></div>`;
 
     const selectionCheckbox = column.id !== 'selectionRadio' && column.id === 'selectionCheckbox';
     const colName = escapeHTML(column.name);
@@ -398,6 +394,7 @@ export default class IdsDataGridHeader extends IdsEventsMixin(IdsElement) {
     const html = `
       <span
         class="ids-data-grid-header-cell${align}${frozen}"
+        ${column.reorderable ? 'draggable="true"' : ''}
         part="header-cell"
         aria-colindex="${index + 1}"
         column-id="${column.id}"