+++ src/components/ids-toolbar/demos/index.ts
@@ -4,3 +4,5 @@ import '../../ids-button/ids-button';
 import '../../ids-input/ids-input';
 import '../../ids-menu-button/ids-menu-button';
 import '../../ids-popup-menu/ids-popup-menu';
+import '../../ids-layout-flex/ids-layout-flex';
+import '../../ids-hyperlink/ids-hyperlink';