+++ test/ids-search-field/ids-search-field-percy-test.ts
@@ -1,4 +1,4 @@
-import percySnapshot from '@percy/puppeteer';
+import pageSnapshot from '../helpers/page-snapshot';
 
 describe('Ids Search Field Percy Tests', () => {
   const url = 'http://localhost:4444/ids-search-field/example.html';
@@ -6,25 +6,25 @@ describe('Ids Search Field Percy Tests', () => {
   it('should not have visual regressions in new light theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['domcontentloaded', 'networkidle0'] });
     await page.waitForSelector('pierce/.ids-input.color-variant-alternate .ids-input-field');
-    await percySnapshot(page, 'ids-search-field-new-light');
+    await pageSnapshot(page, 'ids-search-field-new-light');
   });
 
-  it('should not have visual regressions in new dark theme (percy)', async () => {
+  it.skip('should not have visual regressions in new dark theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['domcontentloaded', 'networkidle0'] });
     await page.evaluate(() => {
-      document.querySelector('ids-theme-switcher')?.setAttribute('mode', 'dark');
+      document.querySelector('ids-theme-switcher')?.setAttribute('theme', 'default-dark');
     });
     await page.waitForSelector('pierce/.ids-input.color-variant-alternate .ids-input-field');
-    await percySnapshot(page, 'ids-search-field-new-dark');
+    await pageSnapshot(page, 'ids-search-field-new-dark');
   });
 
-  it('should not have visual regressions in new contrast theme (percy)', async () => {
+  it.skip('should not have visual regressions in new contrast theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['domcontentloaded', 'networkidle0'] });
     await page.evaluate(() => {
-      document.querySelector('ids-theme-switcher')?.setAttribute('mode', 'contrast');
+      document.querySelector('ids-theme-switcher')?.setAttribute('theme', 'default-contrast');
     });
 
     await page.waitForSelector('pierce/.ids-input.color-variant-alternate .ids-input-field');
-    await percySnapshot(page, 'ids-search-field-new-contrast');
+    await pageSnapshot(page, 'ids-search-field-new-contrast');
   });
 });